requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==1.1.0
google-api-python-client==2.108.0
pytz==2023.3
python-dotenv==1.0.0